% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM2FSC.R
\name{SDM2FSC}
\alias{SDM2FSC}
\title{2-Span Slope Deflection Method(Fixed-Support-Cantilever)}
\usage{
SDM2FSC(LengthAB, LengthBC, Typeof1, Typeof2, Magnitude1, Magnitude2, Pointof1,
  Pointof2, EAB, EBC, IAB, IBC, CM, CMomentDirection)
}
\arguments{
\item{LengthAB}{The length of span AB}

\item{LengthBC}{The length of span BC}

\item{Typeof1}{The type of load(i.e., UDL/PL/TDLS/TDLE) on span AB}

\item{Typeof2}{The type of load(i.e., UDL/PL/TDLS/TDLE) on span BC}

\item{Magnitude1}{The magnitude of the load on span AB}

\item{Magnitude2}{The magnitude of the load on span BC}

\item{Pointof1}{The point at  which the load is applied on span AB}

\item{Pointof2}{The point at  which the load is applied on span BC}

\item{EAB}{Young's Modulus(E) of span AB}

\item{EBC}{Young's Modulus(E) of span BC}

\item{IAB}{Inertia(second moment of area, I) of span AB}

\item{IBC}{Inertia(second moment of area, I) of span BC}

\item{CM}{Moment value}

\item{CMomentDirection}{Moment direction}
}
\value{
Moments and Rotations
}
\description{
This function computes the moments and the rotations using slope deflection method(SDM).
}
\details{
Given the span length, magnitude of load, type of load, point of action, Young's Modulus, and interia, the function computes the moments and the rotations using slope deflection method(SDM).
}
\examples{
Single load on a span (http://www.abzwater.com/sdm)
SDM2FSC(3,2,"UDL","PL",20,40,1,1,1,1,1,1,10,"C")
SDM2FSC(4,8,"PL","PL",0,40,1,4,1,1,1,1,5,"C")
SDM2FSC(2,2,"PL","UDL",40,20,1,1,1,1,1,1,40,"C")


}
\references{
(1) Wikipedia contributors. (2018, February 19). Slope deflection method. In Wikipedia, The Free Encyclopedia. Retrieved 09:17, March 26, 2018, from https://en.wikipedia.org/w/index.php?title=Slope_deflection_method&oldid=826540985
\cr
\cr
(2) Norris, Charles Head; John Benson Wilbur; Senol Utku (1976). Elementary Structural Analysis (3rd ed.). McGraw-Hill. pp. 313–326. ISBN 0-07-047256-4.
\cr
\cr
(3) McCormac, Jack C.; Nelson, James K. Jr. (1997). Structural Analysis: A Classical and Matrix Approach (2nd ed.). Addison-Wesley. pp. 430–451. ISBN 0-673-99753-7.
\cr
\cr
}
\seealso{
\link{SDM2FS}
\link{SDM2FF}
}
\author{
\strong{Sivarajah Mylevaganam}
}
\keyword{Civil}
\keyword{Deflection}
\keyword{Engineering}
\keyword{Indeterminate}
\keyword{Slope}
\keyword{Structural}
