% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM3FF.R
\name{SDM3FF}
\alias{SDM3FF}
\title{3-Span Slope Deflection Method(Fixed-Fixed)}
\usage{
SDM3FF(LengthAB, LengthBC, LengthCD, Typeof1, Typeof2, Typeof3, Magnitude1,
  Magnitude2, Magnitude3, Pointof1, Pointof2, Pointof3, EAB, EBC, ECD, IAB, IBC,
  ICD)
}
\arguments{
\item{LengthAB}{The length of span AB}

\item{LengthBC}{The length of span BC}

\item{LengthCD}{The length of span CD}

\item{Typeof1}{The type of load(i.e., UDL/PL/TDLS/TDLE) on span AB}

\item{Typeof2}{The type of load(i.e., UDL/PL/TDLS/TDLE) on span BC}

\item{Typeof3}{The type of load(i.e., UDL/PL/TDLS/TDLE) on span CD}

\item{Magnitude1}{The magnitude of the load on span AB}

\item{Magnitude2}{The magnitude of the load on span BC}

\item{Magnitude3}{The magnitude of the load on span CD}

\item{Pointof1}{The point at  which the load is applied on span AB}

\item{Pointof2}{The point at  which the load is applied on span BC}

\item{Pointof3}{The point at  which the load is applied on span CD}

\item{EAB}{Young's Modulus(E) of span AB}

\item{EBC}{Young's Modulus(E) of span BC}

\item{ECD}{Young's Modulus(E) of span CD}

\item{IAB}{Inertia(second moment of area, I) of span AB}

\item{IBC}{Inertia(second moment of area, I) of span BC}

\item{ICD}{Inertia(second moment of area, I) of span CD}
}
\value{
Moments and Rotations
}
\description{
This function computes the moments and the rotations using slope deflection method(SDM).
}
\details{
Given the span length, magnitude of load, type of load, point of action, Young's Modulus, and interia, the function computes the moments and the rotations using SDM.
}
\note{
The possible values for Typeof1, Typeof2, and Typeof3 are UDL/PL/TDLS/TDLE.\cr
UDL:Uniformly Distributed Load \cr
PL:Point Load\cr
TDLS:Triangular Distributed Load(the maximum value is at the start, see http://www.abzwater.com/sdm)\cr
TDLE:Triangular Distributed Load(the maximum value is at the end, see http://www.abzwater.com/sdm )\cr
}
\examples{
Single load on a span (http://www.abzwater.com/sdm)
SDM3FF(8,4,8,"PL","UDL","PL",40,24,20,4,1,4,1,1,1,1,1,1)
SDM3FF(2,2,2,"UDL","UDL","UDL",40,40,40,1,1,1,1,1,1,1,1,1)
SDM3FF(2,2,2,"UDL","UDL","UDL",40,80,40,1,1,1,1,1,1,1,1,1)


Multiple load on a span(http://www.abzwater.com/sdm)
SDM3FF(5,10,5,"PL","PL","PL",0,40,0,1,3,1,1,1,1,1,1,1) + SDM3FF(5,10,5,"PL","PL","PL",0,40,0,1,7,1,1,1,1,1,1,1)

Framed  Structure(http://www.abzwater.com/sdm)
SDM3FF(5,10,5,"PL","UDL","PL",0,5,0,1,1,1,1,1,1,1,1,1)



}
\references{
(1) Wikipedia contributors. (2018, February 19). Slope deflection method. In Wikipedia, The Free Encyclopedia. Retrieved 09:17, March 26, 2018, from https://en.wikipedia.org/w/index.php?title=Slope_deflection_method&oldid=826540985
\cr
\cr
(2) Norris, Charles Head; John Benson Wilbur; Senol Utku (1976). Elementary Structural Analysis (3rd ed.). McGraw-Hill. pp. 313–326. ISBN 0-07-047256-4.
\cr
\cr
(3) McCormac, Jack C.; Nelson, James K. Jr. (1997). Structural Analysis: A Classical and Matrix Approach (2nd ed.). Addison-Wesley. pp. 430–451. ISBN 0-673-99753-7.
\cr
\cr
}
\seealso{
\link{SDM2FS}
\link{SDM2FSC}
\link{SDM2FF}
}
\author{
\strong{Sivarajah Mylevaganam}
}
\keyword{Civil}
\keyword{Deflection}
\keyword{Engineering}
\keyword{Indeterminate}
\keyword{Slope}
\keyword{Structural}
